+++ previous-versions/assess/src/sdk/models/operations/getaccountsforenhancedbalancesheet.ts
@@ -55,5 +55,5 @@ export class GetAccountsForEnhancedBalanceSheetResponse extends SpeakeasyBase {
      * Raw HTTP response; suitable for custom response parsing
      */
     @SpeakeasyMetadata()
-    rawResponse?: AxiosResponse;
+    rawResponse: AxiosResponse;
 }