+++ previous-versions/assess/src/sdk/models/operations/getexcelreport.ts
@@ -51,5 +51,5 @@ export class GetExcelReportResponse extends SpeakeasyBase {
      * Raw HTTP response; suitable for custom response parsing
      */
     @SpeakeasyMetadata()
-    rawResponse?: AxiosResponse;
+    rawResponse: AxiosResponse;
 }