+++ previous-versions/assess/src/sdk/sdk.ts
@@ -56,9 +56,9 @@ export class SDKConfiguration {
     serverDefaults: any;
     language = "typescript";
     openapiDocVersion = "1.0";
-    sdkVersion = "0.37.0";
-    genVersion = "2.188.1";
-    userAgent = "speakeasy-sdk/typescript 0.37.0 2.188.1 1.0 @codat/assess";
+    sdkVersion = "0.37.1";
+    genVersion = "2.194.1";
+    userAgent = "speakeasy-sdk/typescript 0.37.1 2.194.1 1.0 @codat/assess";
     retryConfig?: utils.RetryConfig;
     public constructor(init?: Partial<SDKConfiguration>) {
         Object.assign(this, init);