+++ previous-versions/assess/RELEASES.md
@@ -637,3 +637,13 @@ Based on:
 - [typescript v0.37.0] previous-versions/assess
 ### Releases
 - [NPM v0.37.0] https://www.npmjs.com/package/@codat/assess/v/0.37.0 - previous-versions/assess
+
+## 2023-11-20 01:26:01
+### Changes
+Based on:
+- OpenAPI Doc 1.0 https://raw.githubusercontent.com/codatio/oas/main/yaml/Codat-Assess.yaml
+- Speakeasy CLI 1.121.1 (2.194.1) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v0.37.1] previous-versions/assess
+### Releases
+- [NPM v0.37.1] https://www.npmjs.com/package/@codat/assess/v/0.37.1 - previous-versions/assess